Split build graph lowering tests
